requirements.txt
azure-identity
azure-keyvault-secrets
requests
pynacl
python-dotenv
